pyproject.toml
[tool.poetry]
name = "st-backend"
version = "0.1.0"
description = ""
authors = ["Adnan Baysal <adbaysal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
celery = {extras = ["redis"], version = "^5.3.1"}
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
pyyaml = "^6.0.1"
uritemplate = "^4.1.1"
drf-spectacular = "^0.26.4"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
